omie_sync/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
psycopg2-binary